Move game starting logic into manager
