+++ backend/routes/game-api.ts
@@ -3,14 +3,10 @@
  */
 import * as express from 'express';
 
-import {config} from '../config';
 import {BorisDatabase} from '../db/db';
-import { START_GAME, ABANDON_GAME, GET_UI_STATE, GameStatus } from '../../common/api';
+import { START_GAME, ABANDON_GAME, GET_UI_STATE } from '../../common/api';
 import { makeApiHelper, RequireUser, SafeError } from './api-utils';
-import { Scenario } from '../../common/models';
 import { GameManager } from '../game/manager';
-import { publishEvent } from '../websocket/pub-sub';
-import { NotificationType } from '../../common/notifications';
 
 export const router = express.Router();
 
@@ -49,34 +45,8 @@ async function getActiveGameForUser(app: express.Application, userId: number): P
 apiMethod(START_GAME, async (data, app, user) => {
     const db: BorisDatabase = app.get("db");
     const activeTeamMembership = await requireActiveTeamMembership(db, user.id);
-    const team = await db.teams.findOne({id: activeTeamMembership.team_id});
-    const teamMembers = +await db.team_members.count({team_id: team.id, is_active: true});
-    if (teamMembers < 2) {
-        throw new SafeError("You must have at least two people on your team to play.");
-    } else if (teamMembers > 5) {
-        throw new SafeError("Too many people on the team."); // This shouldn't be possible.
-    }
-    const scenario = await db.scenarios.findOne({id: data.scenarioId, is_active: true});
-    if (scenario === null) {
-        throw new SafeError("Invalid scenario.");
-    }
-    // Start the game!
-    try {
-        await db.games.insert({
-            team_id: team.id,
-            scenario_id: scenario.id,
-            is_active: true,
-        });
-    } catch (error) {
-        throw new SafeError("Unable to start playing. Did the game already start?");
-    }
-    const gameStatus: GameStatus = {
-        scenarioId: scenario.id,
-        scenarioName: scenario.name,
-        isActive: true,
-    };
-    publishEvent(team.id, {type: NotificationType.GAME_STATUS_CHANGED, ...gameStatus});
-    return gameStatus;
+    let {status} = await GameManager.startGame(activeTeamMembership.team_id, data.scenarioId);
+    return status;
 });
 
 /**
